requirements.txt
p2pnetwork==0.0.3
pycryptodome==3.10.1
setuptools~=56.0.0
xlrd~=2.0.1
xlwt~=1.3.0
xlutils~=2.0.0
XlsxWriter~=1.4.3